Health-Tracker-Backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
Django==4.0.4
django-bootstrap4==22.1
django-cors-headers==3.11.0
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
psycopg2==2.9.3
PyJWT==1.7.1
pytz==2022.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
